Rework template
