+++ content/visualisation/2021-06-24-Increased-snack-intake-in-UK-and-Australia/index.mdx
@@ -7,9 +7,6 @@ category: ["Department of Psychology"]
 tag: ["COVID-19 lockdown", "Food intake", "Craving control"]
 description: A new study led by the University of Sheffield suggests that treats, such as sweets, chocolate and salty treats, increased for some people during Covid-19 lockdowns. People with low control over their cravings were most at risk of increasing their snack intake during the lockdown. Researchers warn about weight and health related issues for at-risk adults due to the impact of lockdowns on our eating behaviours.
 date: "2021-06-24"
-template:
-featured: false
-isPublished: false
 rowSpan: 1
 columnSpan: 1
 ---
