+++ src/templates/visualisation/visItemTemplate.jsx
@@ -1,0 +1,95 @@
+// BASE
+import React from "react"
+import { graphql, withPrefix } from "gatsby"
+import PropTypes from "prop-types"
+import Helmet from "react-helmet"
+
+import SEO from "../../components/shared/seo"
+import PostPagination from "../../components/blog/postPagination"
+import Comment from "../../components/blog/comment"
+
+// MDX component
+import "katex/dist/katex.min.css"
+import CommonMdxProvider from "../../components/shared/commonMdxProvider"
+
+// Utils
+import { useScript } from "../../utils/hooks/useScript"
+import {  getShareLinks } from "../../utils/shared"
+import VisPagination from "../../components/visualisation/visPagination"
+import VisDetail from "../../components/visualisation/visDetail"
+import { VisDiv } from "../../components/style/visStyle"
+
+
+const visItemTemplate = ({ data: { mdx }, pageContext }) => {
+  const components = { VisPagination, VisDetail, VisDiv };
+
+  const shareLinks = getShareLinks(mdx);
+  const { title, template } = mdx.frontmatter;
+  const {prev, next} = pageContext;
+
+  mdx.shareLinks = shareLinks;
+  mdx.pageContext = pageContext;
+
+  // include d3 scripts
+  const d3 = mdx.frontmatter.d3 || null;
+
+  {d3 && d3.map((d) => {
+    if (d.includes("https://")) {
+      useScript(d, "", false);  // external script
+    } else {
+      useScript(withPrefix(`d3/${d}`), "", false);   
+    }
+  })}
+
+
+  return (
+    <div className="relative" key={mdx.id}>
+      <SEO title={title} keywords={["the university of sheffield", "data visualisation", "data visualisation hub", "research", "blog"]} />
+      <Helmet >
+        <script async src="https://platform.twitter.com/widgets.js" charset="utf-8" type = 'text/javascript' /> 
+      </Helmet>
+      
+      {/* body */}
+      {template === "custom" ? 
+        <div className="justify-center mx-auto text-lg lg:text-xl">
+          <CommonMdxProvider mdx={mdx} components={components} />
+        </div>
+        :
+        <div className="flex flex-wrap relative mx-auto bg-gray-900">
+          <div className="relative mx-auto container py-32 px-5 text-lg xl:text-xl text-gray-200" style={{maxWidth: '1200px'}}>
+            <CommonMdxProvider mdx={mdx} components={components} />
+          </div>
+        </div>   
+      }    
+ 
+      <PostPagination mdx={mdx} prev={prev} next={next} shareLinks={shareLinks} displayPagination={false} displayTags={false} displayShareButtons={false} />
+      <Comment mdx={mdx} />
+    </div>
+  );
+}
+
+export default visItemTemplate
+
+visItemTemplate.propTypes = {
+  data: PropTypes.any,
+  pageContext: PropTypes.any,
+}
+
+export const query = graphql`
+  query visualisationItemQuery($id: String) {
+    mdx(id: {eq: $id}) {
+      id
+      body
+      tableOfContents
+      fields {
+        ...MdxFields
+        slugOrigin
+      }
+      frontmatter {
+        ...MdxFrontmatter
+        disableTOC
+        d3
+      }
+    }
+  }
+`
